orchestrator/requirements.txt
Django==6.0.0
PyYAML==6.0.3
psycopg2-binary==2.9.11